requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
Pillow==10.0.1
django-filter==23.3
dj-database-url==2.1.0
whitenoise==6.6.0
celery==5.3.4
redis==5.0.1
django-environ==0.11.2